Reformat the 'Victory' star to look more aesthetic
